feat(coach): decommission atdd babysit — stub + archive (#532)
- Move commands/babysit.py (~989 lines) to commands/_archived/babysit.py
  for parity-test reuse by the L8 fixture suite (#525)
- Install stub at commands/babysit.py that prints migration message and
  exits non-zero (no babysit machinery executes)
- Update cli.py dispatch: strip all babysit kwargs from the call site,
  routing any invocation of `atdd babysit` to the stub

Absorption map per spec §0.2:
  token-count alerting     → observer rule 06-token-threshold
  bash auto-approval       → observer rule 13-bash-auto-approve
  aggregate approval       → atdd observer aggregate-approve
  naming/layout drift      → observer rules 14/15
  violation detection      → observer rules 04/16
  dashboard                → atdd observer status
  workspace polling        → replaced by event-driven runtime_watcher
  phase-cache-via-labels   → replaced by coach state machine
